correction to TTEntry

+++ src/types.h
@@ -23,7 +23,7 @@
 namespace Engine {
 
 using Bitboard = uint64_t;
-using Depth = uint16_t;
+using Depth = int;
 
 constexpr int MAX_MOVES = 256;
 constexpr int MAX_PLY = 246;
